a couple of quick fixes

+++ TownOfUsReworked/PlayerLayers/Roles/Apocalypse/Deity/Pestilence.cs
@@ -19,7 +19,7 @@ public sealed class Pestilence : Deity
     public override Func<string> Description => () => "- You can spread a deadly disease to other players" + CommonAbilities;
     public override bool CanVent => base.CanVent && PestVent;
 
-    public static readonly Dictionary<byte, uint> Infected = [];
+    public static readonly Dictionary<byte, byte> Infected = [];
 
     protected override void Init()
     {
